Modify few broken Links and Re-design Quizzes Page
Fixed a few broken links are removed urls that
lead to nowhere. Redesigned the Main Quizzes page
by removing the sub menus and instead added filter
funtionality for the user to group and view
quizzes based on their state like 'live' or 'done'
and users interaction with the quiz - attempted or
not.